ea: Change type of array stride parameters to ptrdiff_t
ptrdiff_t is the correct type for array strides and similar.

+++ libavcodec/eaidct.h
@@ -19,8 +19,9 @@
 #ifndef AVCODEC_EAIDCT_H
 #define AVCODEC_EAIDCT_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-void ff_ea_idct_put_c(uint8_t *dest, int linesize, int16_t *block);
+void ff_ea_idct_put_c(uint8_t *dest, ptrdiff_t linesize, int16_t *block);
 
 #endif /* AVCODEC_EAIDCT_H */
